+++ service/braket/api.go
@@ -1508,6 +1508,67 @@ func (s *AlgorithmSpecification) SetScriptModeConfig(v *ScriptModeConfig) *Algor
 	return s
 }
 
+// The Amazon Braket resource and the association type.
+type Association struct {
+	_ struct{} `type:"structure"`
+
+	// The Amazon Braket resource arn.
+	//
+	// Arn is a required field
+	Arn *string `locationName:"arn" type:"string" required:"true"`
+
+	// The association type for the specified Amazon Braket resource arn.
+	//
+	// Type is a required field
+	Type *string `locationName:"type" type:"string" required:"true" enum:"AssociationType"`
+}
+
+// String returns the string representation.
+//
+// API parameter values that are decorated as "sensitive" in the API will not
+// be included in the string output. The member name will be present, but the
+// value will be replaced with "sensitive".
+func (s Association) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation.
+//
+// API parameter values that are decorated as "sensitive" in the API will not
+// be included in the string output. The member name will be present, but the
+// value will be replaced with "sensitive".
+func (s Association) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *Association) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "Association"}
+	if s.Arn == nil {
+		invalidParams.Add(request.NewErrParamRequired("Arn"))
+	}
+	if s.Type == nil {
+		invalidParams.Add(request.NewErrParamRequired("Type"))
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetArn sets the Arn field's value.
+func (s *Association) SetArn(v string) *Association {
+	s.Arn = &v
+	return s
+}
+
+// SetType sets the Type field's value.
+func (s *Association) SetType(v string) *Association {
+	s.Type = &v
+	return s
+}
+
 type CancelJobInput struct {
 	_ struct{} `type:"structure" nopayload:"true"`
 
@@ -1830,6 +1891,9 @@ type CreateJobInput struct {
 	// AlgorithmSpecification is a required field
 	AlgorithmSpecification *AlgorithmSpecification `locationName:"algorithmSpecification" type:"structure" required:"true"`
 
+	// The list of Amazon Braket resources associated with the hybrid job.
+	Associations []*Association `locationName:"associations" type:"list"`
+
 	// Information about the output locations for job checkpoint data.
 	CheckpointConfig *JobCheckpointConfig `locationName:"checkpointConfig" type:"structure"`
 
@@ -1935,6 +1999,16 @@ func (s *CreateJobInput) Validate() error {
 			invalidParams.AddNested("AlgorithmSpecification", err.(request.ErrInvalidParams))
 		}
 	}
+	if s.Associations != nil {
+		for i, v := range s.Associations {
+			if v == nil {
+				continue
+			}
+			if err := v.Validate(); err != nil {
+				invalidParams.AddNested(fmt.Sprintf("%s[%v]", "Associations", i), err.(request.ErrInvalidParams))
+			}
+		}
+	}
 	if s.CheckpointConfig != nil {
 		if err := s.CheckpointConfig.Validate(); err != nil {
 			invalidParams.AddNested("CheckpointConfig", err.(request.ErrInvalidParams))
@@ -1980,6 +2054,12 @@ func (s *CreateJobInput) Validate() error {
 // SetAlgorithmSpecification sets the AlgorithmSpecification field's value.
 func (s *CreateJobInput) SetAlgorithmSpecification(v *AlgorithmSpecification) *CreateJobInput {
 	s.AlgorithmSpecification = v
+	return s
+}
+
+// SetAssociations sets the Associations field's value.
+func (s *CreateJobInput) SetAssociations(v []*Association) *CreateJobInput {
+	s.Associations = v
 	return s
 }
 
@@ -2090,6 +2170,9 @@ type CreateQuantumTaskInput struct {
 	// Action is a required field
 	Action aws.JSONValue `locationName:"action" type:"jsonvalue" required:"true"`
 
+	// The list of Amazon Braket resources associated with the quantum task.
+	Associations []*Association `locationName:"associations" type:"list"`
+
 	// The client token associated with the request.
 	ClientToken *string `locationName:"clientToken" min:"1" type:"string" idempotencyToken:"true"`
 
@@ -2174,6 +2257,16 @@ func (s *CreateQuantumTaskInput) Validate() error {
 	if s.Shots == nil {
 		invalidParams.Add(request.NewErrParamRequired("Shots"))
 	}
+	if s.Associations != nil {
+		for i, v := range s.Associations {
+			if v == nil {
+				continue
+			}
+			if err := v.Validate(); err != nil {
+				invalidParams.AddNested(fmt.Sprintf("%s[%v]", "Associations", i), err.(request.ErrInvalidParams))
+			}
+		}
+	}
 
 	if invalidParams.Len() > 0 {
 		return invalidParams
@@ -2184,6 +2277,12 @@ func (s *CreateQuantumTaskInput) Validate() error {
 // SetAction sets the Action field's value.
 func (s *CreateQuantumTaskInput) SetAction(v aws.JSONValue) *CreateQuantumTaskInput {
 	s.Action = v
+	return s
+}
+
+// SetAssociations sets the Associations field's value.
+func (s *CreateQuantumTaskInput) SetAssociations(v []*Association) *CreateQuantumTaskInput {
+	s.Associations = v
 	return s
 }
 
@@ -2848,6 +2947,9 @@ type GetJobOutput struct {
 	// AlgorithmSpecification is a required field
 	AlgorithmSpecification *AlgorithmSpecification `locationName:"algorithmSpecification" type:"structure" required:"true"`
 
+	// The list of Amazon Braket resources associated with the hybrid job.
+	Associations []*Association `locationName:"associations" type:"list"`
+
 	// The billable time the Amazon Braket job used to complete.
 	BillableDuration *int64 `locationName:"billableDuration" type:"integer"`
 
@@ -2953,6 +3055,12 @@ func (s GetJobOutput) GoString() string {
 // SetAlgorithmSpecification sets the AlgorithmSpecification field's value.
 func (s *GetJobOutput) SetAlgorithmSpecification(v *AlgorithmSpecification) *GetJobOutput {
 	s.AlgorithmSpecification = v
+	return s
+}
+
+// SetAssociations sets the Associations field's value.
+func (s *GetJobOutput) SetAssociations(v []*Association) *GetJobOutput {
+	s.Associations = v
 	return s
 }
 
@@ -3076,7 +3184,7 @@ type GetQuantumTaskInput struct {
 	// A list of attributes to return information for.
 	AdditionalAttributeNames []*string `location:"querystring" locationName:"additionalAttributeNames" type:"list" enum:"QuantumTaskAdditionalAttributeName"`
 
-	// the ARN of the task to retrieve.
+	// The ARN of the task to retrieve.
 	//
 	// QuantumTaskArn is a required field
 	QuantumTaskArn *string `location:"uri" locationName:"quantumTaskArn" type:"string" required:"true"`
@@ -3130,6 +3238,9 @@ func (s *GetQuantumTaskInput) SetQuantumTaskArn(v string) *GetQuantumTaskInput {
 
 type GetQuantumTaskOutput struct {
 	_ struct{} `type:"structure"`
+
+	// The list of Amazon Braket resources associated with the quantum task.
+	Associations []*Association `locationName:"associations" type:"list"`
 
 	// The time at which the task was created.
 	//
@@ -3205,6 +3316,12 @@ func (s GetQuantumTaskOutput) String() string {
 // value will be replaced with "sensitive".
 func (s GetQuantumTaskOutput) GoString() string {
 	return s.String()
+}
+
+// SetAssociations sets the Associations field's value.
+func (s *GetQuantumTaskOutput) SetAssociations(v []*Association) *GetQuantumTaskOutput {
+	s.Associations = v
+	return s
 }
 
 // SetCreatedAt sets the CreatedAt field's value.
@@ -3640,7 +3757,7 @@ type JobEventDetails struct {
 	// job.
 	Message *string `locationName:"message" type:"string"`
 
-	// TThe type of event that occurred related to the Amazon Braket job.
+	// The type of event that occurred related to the Amazon Braket job.
 	TimeOfEvent *time.Time `locationName:"timeOfEvent" type:"timestamp" timestampFormat:"iso8601"`
 }
 
@@ -5305,6 +5422,18 @@ func (s *ValidationException) StatusCode() int {
 // RequestID returns the service's response RequestID for request.
 func (s *ValidationException) RequestID() string {
 	return s.RespMetadata.RequestID
+}
+
+const (
+	// AssociationTypeReservationTimeWindowArn is a AssociationType enum value
+	AssociationTypeReservationTimeWindowArn = "RESERVATION_TIME_WINDOW_ARN"
+)
+
+// AssociationType_Values returns all elements of the AssociationType enum
+func AssociationType_Values() []string {
+	return []string{
+		AssociationTypeReservationTimeWindowArn,
+	}
 }
 
 const (